requirements.txt
asgiref==3.4.1
astroid==2.8.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
PyJWT==2.1.0
pylint==2.11.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1
